// server/types.ts
export enum Role {
  DIRECTOR = 'DIRECTOR',
  ACCOUNTANT = 'ACCOUNTANT',
  AGENT = 'AGENT',
  CLIENT = 'CLIENT'
}

export enum ShipmentStatus {
  OPENED = 'Ouverture Dossier',
  PRE_CLEARANCE = 'Pré-Dédouanement (DDI & BSC)', // Étape 1 stricte
  CUSTOMS_LIQUIDATION = 'Liquidation Douane',
  LIQUIDATION_PAID = 'Liquidation Payée', // Étape 4 : Paiement validé
  BAE_GRANTED = 'BAE Obtenu',
  PORT_EXIT = 'Sortie Port',
  DELIVERED = 'Livré / Archivé'
}